requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
alembic==1.13.3
psycopg[binary]==3.2.3
kafka-python==2.0.2
pytest==8.3.3
httpx==0.27.2
